회원가입, 로그인, 작성자 닉네임, 조회수 구현

+++ src/main/java/com/yooboong/board/controller/CommentController.java
@@ -3,9 +3,14 @@ package com.yooboong.board.controller;
 import com.yooboong.board.dto.CommentDto;
 import com.yooboong.board.service.CommentService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
+
+import java.security.Principal;
 
 @Controller
 @RequiredArgsConstructor
@@ -14,44 +19,63 @@ public class CommentController {
 
     private final CommentService commentService;
 
+    @PreAuthorize("isAuthenticated()")
     @PostMapping("/create")
-    public String create(@RequestParam("postingId") Long postingId,
+    public String create(Principal principal,
+                         @RequestParam("postingId") Long postingId,
                          @RequestParam("comment") String comment) {
         CommentDto input = CommentDto.builder()
                 .postingId(postingId)
                 .comment(comment)
                 .build();
 
-        CommentDto created = commentService.create(postingId, input);
+        CommentDto created = commentService.create(principal.getName(), postingId, input);
         return "redirect:/posting/" + postingId;
     }
 
+    @PreAuthorize("isAuthenticated()")
     @GetMapping("/{id}/edit")
-    public String edit(@PathVariable("id") Long id,
-                       Model model) {
-        CommentDto commentDto = commentService.read(id);
+    public String editForm(Principal principal,
+                           @PathVariable("id") Long id,
+                           Model model) {
+        CommentDto target = commentService.read(id);
 
-        model.addAttribute("commentDto", commentDto);
+        if (!target.getUsername().equals(principal.getName()))
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "댓글 수정 권한이 없음");
+
+        model.addAttribute("commentDto", target);
         return "comments/edit";
     }
 
+    @PreAuthorize("isAuthenticated()")
     @PostMapping("/update")
-    public String update(@RequestParam("id") Long id,
+    public String update(Principal principal,
+                         @RequestParam("id") Long id,
                          @RequestParam("comment") String comment) {
-        CommentDto commentDto = CommentDto.builder()
+        CommentDto target = commentService.read(id);
+        if (!target.getUsername().equals(principal.getName()))
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "댓글 수정 권한이 없음");
+
+        CommentDto input = CommentDto.builder()
                 .id(id)
                 .comment(comment)
                 .build();
 
-        CommentDto updated = commentService.update(id, commentDto);
+        CommentDto updated = commentService.update(id, input);
 
         Long postingId = updated.getPostingId();
 
         return "redirect:/posting/" + postingId;
     }
 
+    @PreAuthorize("isAuthenticated()")
     @GetMapping("/{id}/delete")
-    public String delete(@PathVariable("id") Long id) {
+    public String delete(Principal principal,
+                         @PathVariable("id") Long id) {
+        CommentDto target = commentService.read(id);
+        if (!target.getUsername().equals(principal.getName()))
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "댓글 삭제 권한이 없음");
+
         Long deletedCommentPostingId = commentService.delete(id);
 
         return "redirect:/posting/" + deletedCommentPostingId;
