+++ src/main/java/com/yooboong/board/controller/PostingController.java
@@ -6,10 +6,14 @@ import com.yooboong.board.service.CommentService;
 import com.yooboong.board.service.PostingService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
+import java.security.Principal;
 import java.util.List;
 
 @Controller
@@ -41,25 +45,29 @@ public class PostingController {
         model.addAttribute("currentPage", currentPage);
         model.addAttribute("lastPage", lastPage);
         model.addAttribute("searchOption", searchOption);
-
         return "mainpage";
     }
 
+    @PreAuthorize("isAuthenticated()")
     @GetMapping("/new") // 글 생성 버튼 눌렀을 때
     public String createForm() {
         return "posting/new"; // 글 생성 페이지로 이동
     }
 
+    @PreAuthorize("isAuthenticated()") // 로그인한 경우에만 실행, 로그인한 사용자만 호출가능
+    // 로그아웃 상태에서 해당 어노테이션이 적용된 메소드가 호출되면, 로그인 페이지로 강제이동
     @PostMapping("/create") // 글 생성
-    public String create(@RequestParam("title") String title,
+    public String create(Principal principal,
+                         @RequestParam("title") String title,
                          @RequestParam("content") String content) {
 
         PostingDto input = PostingDto.builder()
                 .title(title)
                 .content(content)
+                .view(0)
                 .build();
 
-        PostingDto created = postingService.create(input);
+        PostingDto created = postingService.create(principal.getName(), input);
 
         return "redirect:/posting/";
     }
@@ -69,24 +77,37 @@ public class PostingController {
                        Model model) {
         PostingDto postingDto = postingService.read(id);
 
+        postingService.increaseView(id);
+
         model.addAttribute("postingDto", postingDto);
         return "posting/show";
     }
 
+    @PreAuthorize("isAuthenticated()")
     @GetMapping("/{id}/edit") // 수정 시작
-    public String editForm(@PathVariable("id") Long id,
+    public String editForm(Principal principal,
+                           @PathVariable("id") Long id,
                            Model model) {
         // 수정할 데이터 가져오기
         PostingDto target = postingService.read(id);
+
+        if (!target.getUsername().equals(principal.getName()))
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "게시글 수정 권한이 없음");
 
         model.addAttribute("postingDto", target);
         return "/posting/edit";
     }
 
+    @PreAuthorize("isAuthenticated()")
     @PostMapping("/update") // 수정
-    public String update(@RequestParam("id") Long id,
+    public String update(Principal principal,
+                         @RequestParam("id") Long id,
                          @RequestParam("title") String title,
                          @RequestParam("content") String content) {
+        PostingDto target = postingService.read(id);
+        if (!target.getUsername().equals(principal.getName()))
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "게시글 수정 권한이 없음");
+
         PostingDto input = PostingDto.builder()
                 .id(id)
                 .title(title)
@@ -98,8 +119,14 @@ public class PostingController {
         return "redirect:/posting/" + id;
     }
 
+    @PreAuthorize("isAuthenticated()")
     @GetMapping("/{id}/delete") // 삭제
-    public String delete(@PathVariable("id") Long id) {
+    public String delete(Principal principal,
+                         @PathVariable("id") Long id) {
+        PostingDto target = postingService.read(id);
+        if (!target.getUsername().equals(principal.getName()))
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "게시글 삭제 권한이 없음");
+
         postingService.delete(id);
 
         return "redirect:/posting/";
